fix: 채팅 내 메세지, 남의 메세지로 수정
