+++ client/src/components/chat/chat.tsx
@@ -1,20 +1,67 @@
 "use client";
 
-import Message from "./message/message";
 import style from "./chat.module.scss";
 import MessageInput from "./messageInput/messageInput";
 import useChat from "@/hooks/useChat";
 import { useEffect, useRef } from "react";
+import ReceivedMessage from "./receivedMessage/receivedMessage";
+import MyMessage from "./myMessage/myMessage";
 
 interface IChatProps {
     className: string;
     toggleChat: (shouldClose?: boolean) => void;
 }
 
+const TEST_MESSAGE_LIST = [
+    {
+        nickname: "현섭",
+        message: "하이",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "me",
+        message: "dsfdssffsddfdsfdsfdfsdfsfdsdfsfsdfsdfdsfsdf",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "me",
+        message:
+            "ㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋㅋ",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "현섭",
+        message: "ddfjsdhfisdhfsdhuifshduifhsuifhsudifhsiudfhsifhiusfhsdhfuisdfh",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "me",
+        message: "하이",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "현섭",
+        message: "하이",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "현섭",
+        message:
+            "하이이현섭 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+    {
+        nickname: "현섭",
+        message:
+            "하이이현섭 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이이현섭하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭하이이현섭 하이이현섭 하이이현섭",
+        time: "2023-11-07T12:00:00.000Z",
+    },
+];
+
 const Chat = (props: IChatProps) => {
     const { className, toggleChat } = props;
 
-    const { messages } = useChat();
+    // const { messages } = useChat();
 
     const messagesEndRef = useRef<HTMLDivElement>(null);
 
@@ -24,9 +71,9 @@ const Chat = (props: IChatProps) => {
         }
     };
 
-    useEffect(() => {
-        scrollToBottom();
-    }, [messages]);
+    // useEffect(() => {
+    //     scrollToBottom();
+    // }, [messages]);
 
     return (
         <div className={`${className} ${style.chat_container}`}>
@@ -35,14 +82,22 @@ const Chat = (props: IChatProps) => {
                 <button className={style.close_button} onClick={() => toggleChat(true)} />
             </div>
             <div className={style.chat_main}>
-                {messages.map((message, index) => (
-                    <Message
-                        key={index}
-                        message={message.message}
-                        nickname={message.nickname}
-                        time={message.time}
-                    />
-                ))}
+                {TEST_MESSAGE_LIST.map((message, index) => {
+                    return message.nickname === "me" ? (
+                        <ReceivedMessage
+                            key={index}
+                            message={message.message}
+                            time={message.time}
+                        />
+                    ) : (
+                        <MyMessage
+                            key={index}
+                            message={message.message}
+                            nickname={message.nickname}
+                            time={message.time}
+                        />
+                    );
+                })}
                 <div ref={messagesEndRef} />
             </div>
             <div className={style.chat_bottom}>
