+++ client/src/components/chat/receivedMessage/receivedMessage.module.scss
@@ -1,0 +1,41 @@
+@import "../../../styles/variables";
+
+.container {
+    display: flex;
+    gap: 7px;
+    margin-left: auto;
+}
+
+.wrapper {
+    width: 100%;
+    display: flex;
+    flex-flow: column nowrap;
+}
+
+.message_wrapper {
+    display: flex;
+    gap: 6px;
+    margin-top: 6px;
+}
+
+.message_main {
+    padding: 5px 10px;
+    background: rgba(161, 187, 238, 1);
+    border-radius: 15px;
+    font-size: 13px;
+    font-weight: 400;
+    line-height: normal;
+    text-align: left;
+    max-width: 205px;
+    overflow-wrap: break-word;
+    letter-spacing: -0.65px;
+}
+
+.time {
+    display: flex;
+    align-items: flex-end;
+    font-size: 8px;
+    font-weight: 400;
+    color: #c4c4c4;
+    line-height: 13px;
+}
